Rename renderFile to htmlRenderer.
Unlike as a template engine, as middleware vain can get access to
request context, something important for the snippet code to have access
to. They'll be able to examine cookies, request path, session data, etc.

As such, this commit rewires the htmlRenderer to be just a passthru for
reading and returning an HTML file. Names are changed and specs updated
accordingly.