Refactor vain.render to be callback-first.
The template engine I first modeled vain against was blocking-first,
wherein if you provided a callback it resulted in the function doing a
blocking call of itself, then invoked the callback with the result. The
render method is now callback first. If you don't provide a callback
function it will use a Future provided by fibers to return the result
you requested.